#11 - 동작 기록하기
동작을 저장해두기 위해 App.js의 history배열에 기록
Board 컴포넌트의 함수들을 부모 컴포넌트 App로 이동
